Moving endpoints to separate package
