+++ cmd/picow-led-server/endpoints/static.go
@@ -1,13 +1,14 @@
-package main
+package endpoints
 
 import (
 	"net/http"
 
-	"github.com/knackwurstking/picow-led-server/frontend"
 	"github.com/labstack/echo/v4"
+
+	"github.com/knackwurstking/picow-led-server/frontend"
 )
 
-func endpointsStatic() {
+func createStaticEndpoints(e *echo.Echo) {
 	fs := frontend.GetFS()
 	e.StaticFS("/", fs)
 	e.GET("/", func(c echo.Context) error {
