+++ cmd/picow-led-server/endpoints/utils.go
@@ -1,9 +1,8 @@
-package main
+package endpoints
 
 import (
 	"encoding/json"
 	"io"
-	"log/slog"
 	"net/http"
 
 	"github.com/labstack/echo/v4"
@@ -21,15 +20,4 @@ func readBodyData(c echo.Context, data any) (status int, err error) {
 	}
 
 	return http.StatusOK, nil
-}
-
-func saveConfig() {
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		if err := config.save(); err != nil {
-			slog.Warn("Save config", "config.Path", config.Path, "err", err)
-		}
-	}()
-	go config.save()
 }
